refactor: better structure in debugger
