+++ core/src/prelude.rs
@@ -3,6 +3,7 @@ pub use crate::controller::Controller;
 pub use crate::dmg::Dmg;
 pub use crate::joypad::{Joypad, JoypadButton};
 pub use crate::memory::{Accessible, Accessible16};
-pub use crate::ppu::{DMG_SCREEN_HEIGHT, DMG_SCREEN_WIDTH};
+pub use crate::ppu::{DMG_SCREEN_HEIGHT, DMG_SCREEN_WIDTH, Ppu, Renderer};
+pub use crate::serial::SerialListener;
 pub use crate::utils::{self, InstructionBox};
 pub use crate::{bit_accessors, controller, field_bit_accessors, flag_methods, mem_range, reg16};
