+++ frontends/debugger/src/scenes/mod.rs
@@ -1,0 +1,13 @@
+mod emulation;
+mod waiting_file;
+
+pub use emulation::EmulationScene;
+pub use waiting_file::{WaitingFileScene, WaitingFileEvent};
+
+#[derive(Default)]
+pub enum EmulatorState {
+    #[default]
+    Exit,
+    WaitingFile(WaitingFileScene),
+    Emulation(EmulationScene),
+}
